+++ src/bases/DssDirectDepositPoolBase.sol
@@ -45,7 +45,7 @@ interface DssDirectDepositPlanLike {
 
 abstract contract DssDirectDepositPoolBase {
     // --- Auth ---
-    mapping (address => uint) public wards;
+    mapping (address => uint256) public wards;
     function rely(address usr) external auth {
         wards[usr] = 1;
         emit Rely(usr);
